+++ TradableAsset/test/pcalls/Transfer/Transfer.cs
@@ -19,7 +19,7 @@ namespace Expload.Standarts {
             // Emit the asset
             Bytes externalId = new Bytes("0000000000000000000000000000000000000000000000000000000000000001");
             Bytes metaId = new Bytes("0000000000000000000000000000000000000000000000000000000000000002");
-            UInt32 assetId = ProgramHelper.Program<TradableAsset>(programOwner).EmitGTAsset(
+            long assetId = ProgramHelper.Program<TradableAsset>(programOwner).EmitGTAsset(
                 assetOwner, externalId, metaId
             );
